requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
snowflake-connector-python>=3.0.0
snowflake-snowpark-python>=1.6.0
jinja2>=3.1.2
python-dotenv>=1.0.0
requests>=2.28.0
pytest>=7.3.1
httpx>=0.24.0
